gcc13 : fix argument format issues
gcc13 based build failed with below errors.
error: static assertion failed: Cannot format an argument.
  To make type T formattable provide a formatter<T> specialization:
  https://fmt.dev/latest/api.html#udt

Problem reported during enum type input is similar to
https://github.com/fmtlib/fmt/issues/391
Fix is to convert the "enum" type to "int" in the fmt::format function.

Problem related file descriptor input is to convert type to int.

+++ host-transport-extensions/pldm/oem/ibm/pldm_oem_cmds.cpp
@@ -180,7 +180,7 @@ void requestDelete(uint32_t dumpId, uint32_t dumpType)
             fmt::format("Failed to encode pldm FileAck to delete host "
                         "dump,SRC_DUMP_ID({}), "
                         "PLDM_FILE_IO_TYPE({}),PLDM_RETURN_CODE({})",
-                        dumpId, pldmDumpType, retCode)
+                        dumpId, static_cast<int>(pldmDumpType), retCode)
                 .c_str());
         elog<NotAllowed>(Reason("Host dump deletion via pldm is not "
                                 "allowed due to encode fileack failed"));
@@ -197,8 +197,8 @@ void requestDelete(uint32_t dumpId, uint32_t dumpType)
             fmt::format("Failed to send pldm FileAck to delete host dump, "
                         "SRC_DUMP_ID({}), PLDM_FILE_IO_TYPE({}), "
                         "PLDM_RETURN_CODE({}), ERRNO({}), ERRMSG({})",
-                        dumpId, pldmDumpType, retCode, errorNumber,
-                        strerror(errorNumber))
+                        dumpId, static_cast<int>(pldmDumpType), retCode,
+                        errorNumber, strerror(errorNumber))
                 .c_str());
         elog<NotAllowed>(Reason("Host dump deletion via pldm is not "
                                 "allowed due to fileack send failed"));
